Add PhysicalOptimizer to apply rule sets for phsical plans
