+++ sql/core/src/main/scala/org/apache/spark/sql/execution/QueryExecution.scala
@@ -28,9 +28,7 @@ import org.apache.spark.sql.catalyst.analysis.UnsupportedOperationChecker
 import org.apache.spark.sql.catalyst.plans.logical.{LogicalPlan, ReturnAnswer}
 import org.apache.spark.sql.catalyst.rules.RuleExecutor
 import org.apache.spark.sql.catalyst.util.DateTimeUtils
-import org.apache.spark.sql.execution.aggregate.MergePartialAggregate
 import org.apache.spark.sql.execution.command.{DescribeTableCommand, ExecutedCommandExec, ShowTablesCommand}
-import org.apache.spark.sql.execution.exchange.{EnsureRequirements, ReuseExchange}
 import org.apache.spark.sql.types.{BinaryType, DateType, DecimalType, TimestampType, _}
 import org.apache.spark.util.Utils
 
@@ -41,8 +39,7 @@ import org.apache.spark.util.Utils
  * While this is not a public class, we should avoid changing the function names for the sake of
  * changing them, because a lot of developers use the feature for debugging.
  */
-class QueryExecution(val sparkSession: SparkSession, val logical: LogicalPlan)
-    extends RuleExecutor[SparkPlan] {
+class QueryExecution(val sparkSession: SparkSession, val logical: LogicalPlan) {
 
   // TODO: Move the planner an optimizer into here from SessionState.
   protected def planner = sparkSession.sessionState.planner
@@ -87,42 +84,20 @@ class QueryExecution(val sparkSession: SparkSession, val logical: LogicalPlan)
     planner.plan(ReturnAnswer(optimizedPlan)).next()
   }
 
+  /**
+   * Physical query plan optimizer.
+   */
+  lazy val optimizer: RuleExecutor[SparkPlan] = new PhysicalOptimizer(sparkSession)
+
   // executedPlan should not be used to initialize any SparkPlan. It should be
   // only used for execution.
-  lazy val executedPlan: SparkPlan = prepareForExecution(sparkPlan)
+  lazy val executedPlan: SparkPlan = optimizer.execute(sparkPlan)
 
   /** Internal version of the RDD. Avoids copies and has no schema */
   lazy val toRdd: RDD[InternalRow] = executedPlan.execute()
 
-  /**
-   * Prepares a planned [[SparkPlan]] for execution by inserting shuffle operations and internal
-   * row format conversions as needed.
-   */
-  protected def prepareForExecution(plan: SparkPlan): SparkPlan = execute(plan)
-
-  private val fixedPoint = FixedPoint(sparkSession.sessionState.conf.optimizerMaxIterations)
-
-  /** A sequence of rules that will be applied in order to the physical plan before execution. */
-  override protected def batches: Seq[Batch] = Seq(
-    Batch("ExtractPythonUDFs", Once,
-      python.ExtractPythonUDFs),
-    Batch("PlanSubqueries", Once,
-      PlanSubqueries(sparkSession)),
-    Batch("EnsureRequirements", Once,
-      EnsureRequirements(sparkSession.sessionState.conf)),
-    Batch("MergePartialAggregate", fixedPoint,
-      MergePartialAggregate),
-    Batch("CollapseCodegenStages", Once,
-      CollapseCodegenStages(sparkSession.sessionState.conf)),
-    Batch("ReuseResources", Once,
-      ReuseExchange(sparkSession.sessionState.conf),
-      ReuseSubquery(sparkSession.sessionState.conf)
-    )
-  )
-
   protected def stringOrError[A](f: => A): String =
     try f.toString catch { case e: AnalysisException => e.toString }
-
 
   /**
    * Returns the result as a hive compatible sequence of strings. This is for testing only.
